Factorial

+++ Factorial/Factorial_Recursion.c
@@ -1,6 +1,6 @@
 #include<stdio.h>
 
-int factorial(int n){
+long int factorial(int n){
     if(n==0 || n==1) return 1;
     return n*factorial(n-1);
 }
@@ -9,5 +9,5 @@ int main(){
     int a;
     printf("Enter a Number ");
     scanf("%d", &a);
-    printf("The factorial of %d is %d", a, factorial(a));
+    printf("The factorial of %d is %ld", a, factorial(a));
 }
